+++ include/capnp/odTrip.capnp
@@ -24,11 +24,11 @@ struct OdTrip {
   destinationLatitude         @16 :Int32; # divide by 1000000 to get float
   destinationLongitude        @17 :Int32; # divide by 1000000 to get float
   originNodesIdx              @18 :List(Int32);
-  originNodesTravelTimes      @19 :List(Int32); # seconds
-  originNodesDistances        @20 :List(Int32); # meters
+  originNodesTravelTimes      @19 :List(Int16); # seconds
+  originNodesDistances        @20 :List(Int16); # meters
   destinationNodesIdx         @21 :List(Int32);
-  destinationNodesTravelTimes @22 :List(Int32); # seconds
-  destinationNodesDistances   @23 :List(Int32); # meters
+  destinationNodesTravelTimes @22 :List(Int16); # seconds
+  destinationNodesDistances   @23 :List(Int16); # meters
   internalId                  @24 :Text;
   data                        @25 :Text; # json
